Added a patch to tock to support newer nrf52840dk boards
This patch should allow new nordic nrf52840 boards to work with TockOS.
Specifically, without this patch at the first reset of the board, the
approtect mechanism is re-enabled, and therefore it could prevent debugging
with the Segger RTT debugger.

This patch has been inspired by newer versions of TockOS, where the
disabling of the approtect is performed even for the newer versions of
the board. Specifically, I am referring to the following TockOS commits:
- 6d89b4690297bfa6fc53bfb00b7f6f62a0798aed
- 4c88df6123d74ca862bea758cd22fadea32cbb30
- 6d6f43ccf317bd92e988d0bdf6dbd4c550a40938
- 325d1107eb308a33617b84f984997938d84bad2d
- f17ec235dbce883f15c3a06f0d8a1d57312f32fb
- 893a67496c5c21ec53ee92e3db40e7068f2d0f21
- 19f107fdac317643c790e303a96068b4910c6f89
- 2200560f17c4ff93710d5614499368d30834d9ea
- 73780c2500ae74f386d477b8d00ab4b3d65c8772